Add support for building HLSL assets.
Note that this still does not leave output ready to execute.  For that,
I will add an asset packaging step to the build to make it possible to
send binaries over to Stadia instances for execution.

Additionally, this offers no support for an external DXC.  It relies on
the DXC sources checked out in third_party/dxc.  I will add support for
an external DXC toolchain in a subsequent PR.