Fix queries on DXIIVK target
The DXIIVK path should not use GPU queries. They were previously wrapped
in a #ifdef, but a recent CL moved the ifdef such that the queries were
still being created even if the macro was not set. This CL moves the
ifdefs back to their original locations, and also renames it to make it
clear that it should apply to timeline queries also, not just pipeline
queries.