Fluid dynamics simulation.  Initial support. (#136)
Fluid dynamics simulation.  Initial support.

This is still a work in progress.  It is an adaptation of the original
WebGL simulation at https://github.com/PavelDoGreat/WebGL-Fluid-Simulation.

This initial commit adds around 20 compute shaders converted from the
original pixel shaders in WebGL-Fluid-Simulation.

The main logic is not yet implemented.  Rather, this implements a very
basic "bouncing dvd logo" for each compute shader.  Every shader acts on
the same original image, so the result is a bit of an eyesore.

Each compute shader is an instance of a FilterShader class.  In the
future, I will probably add sub-classes for each shader as some of them
have different attributes.

NOTE: I'm including the original GLSL code for reference.  I will remove
it once the application is ready.