Swapchain image count fix for Vulkan (#139)
This is a new one. For some reason, on some Vulkan drivers on
some hardware - the swapchain's image count comes back different
than the minimum image count that's in the create info. This
isn't totally consistent across the driver - meaning that the
same driver on different hardware will not exhibit this behavior.

- grfx::Swapchain::Create now readjusts the mCreateInfo.imageCount
after the API creates the swapchain object and color images.