Fixed FishTornado cbuffer alignment (#141)
The predPos field in FlockingData was incorrectly padded on the
C++ side and was causing a NaN to apper in the first component
of camPos every now and then. This caused the fish flocking to
produce NaNs in the simulation. Most of the time this didn't
happen because memory was somehow zero. This should fix the
issue with the fish not rendering in between runs on Linux and
GGP.

- Corrected alignment of predPos and camPos in FlockingData.

Co-authored-by: Hai Nguyen <haing@google.com>