Add runtime tests to the workflow. (#267)
In order to make `BigWheels` easier to contribute, there should be some
presubmit checks running on the new contributions. This PR uses LavaPipe
from Mesa Vulkan drivers as the software renderer. It adds another step
to the current workflow to exectue some of the `BigWheels` benchmarks as
runtime tests after the code is built and unit tested. `xvfb` is the
display server to run the tests in the headless mode.

Instead of running all benchmarks, we run a subset of them to avoid
duplicate checks and have a faster CI. `square_textured` is used to
check trivial failures such as missing assets. `vk_09_obj_geometry`
and `vk_13_normal_map` are built on top of their previous benchmarks
(`vk_01` to `vk_08`). Hence, running them provides a good coverage.
`Fishtornado` uses a lot of features and assets which makes it a
suitable exhustive test.